% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1_graph.R
\name{eye}
\alias{eye}
\title{function to reproduce the behavior of eye function in matlab}
\usage{
eye(n, m)
}
\arguments{
\item{n}{number of rows in the returned eye matrix}

\item{m}{number of columns in the returned eye matrix}
}
\value{
a matrix with diagonal element as 1 while other elements as zero (eye matrix)
}
\description{
function to reproduce the behavior of eye function in matlab
}
