% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1_graph.R
\name{get_mst}
\alias{get_mst}
\title{function to find the minimum spanning tree}
\usage{
get_mst(X)
}
\arguments{
\item{X}{number of rows in the returned eye matrix (D * N)}
}
\value{
a matrix
}
\description{
function to find the minimum spanning tree
}
