% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_graph_large.R
\name{principal_graph_large}
\alias{principal_graph_large}
\title{function to automatically learn the structure of data by either using L1-graph or the spanning-tree formulization}
\usage{
principal_graph_large(X, y, maxiter = 10, eps = 1e-05,
  gstruct = c("l1-graph", "span-tree"), lambda = 1, gamma = 0.5,
  sigma = 0.01, nn = 5, ncenter = NULL, verbose = T)
}
\arguments{
\item{X}{the input data DxN}

\item{y}{the initial cluster assignment}

\item{maxiter}{maximum number of iteraction}

\item{eps}{relative objective difference}

\item{gstruct}{graph structure to learn, either L1-graph or the span-tree}

\item{lambda}{regularization parameter for inverse graph embedding}

\item{gamma}{regularization parameter for k-means (the prefix of 'param' is used to avoid name collision with gamma)}

\item{sigma}{bandwidth parameter}

\item{nn}{number of nearest neighbors}

\item{verbose}{emit results from iteraction}

\item{C0}{the initialization of centroids}

\item{G}{graph matrix with side information where cannot-link pair is 0}
}
\value{
a list of X, C, W, P, objs
X is the input data
C is the centers for principal graph
W is the pricipal graph matrix
P is the cluster assignment matrix
objs is the objective value for the function
}
\description{
function to automatically learn the structure of data by either using L1-graph or the spanning-tree formulization
}

