% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1_graph.R
\name{get_mst_with_shortcuts}
\alias{get_mst_with_shortcuts}
\title{function to find the minimum spanning tree,
augmented with shortcuts on vertices of odd degree}
\usage{
get_mst_with_shortcuts(X, K = 5)
}
\arguments{
\item{X}{number of rows in the returned eye matrix (D * N)}

\item{K}{of nearest neighbors for use with shortcutting}
}
\value{
a matrix
}
\description{
function to find the minimum spanning tree,
augmented with shortcuts on vertices of odd degree
}
